relocate macros to preprocessor
