+++ ColorzCore/Preprocessor/Macros/IMacro.cs
@@ -1,12 +1,13 @@
 ﻿using ColorzCore.DataTypes;
 using ColorzCore.Lexer;
+using ColorzCore.Parser;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace ColorzCore.Parser.Macros
+namespace ColorzCore.Preprocessor.Macros
 {
     public interface IMacro
     {
